+++ packages/flutter_riverpod/test/providers/change_notifier/change_notifier_provider_test.dart
@@ -91,6 +91,29 @@ void main() {
       );
       expect(root.getAllProviderElements(), isEmpty);
     });
+
+    test('when using provider.overrideWithProvider', () {
+      final provider = ChangeNotifierProvider((ref) => ValueNotifier(0));
+      final root = createContainer();
+      final container = createContainer(parent: root, overrides: [
+        provider.overrideWithProvider(
+          ChangeNotifierProvider((ref) => ValueNotifier(42)),
+        ),
+      ]);
+
+      expect(container.read(provider.notifier).value, 42);
+      expect(container.read(provider).value, 42);
+      expect(
+        container.getAllProviderElements(),
+        unorderedEquals(<Object>[
+          isA<ProviderElementBase>()
+              .having((e) => e.origin, 'origin', provider),
+          isA<ProviderElementBase>()
+              .having((e) => e.origin, 'origin', provider.notifier)
+        ]),
+      );
+      expect(root.getAllProviderElements(), isEmpty);
+    });
   });
 
   test('overriding with value listens to the ChangeNotifier', () {
@@ -284,6 +307,52 @@ void main() {
     expect(container.read(provider.notifier).value, 42);
 
     expect(root.getAllProviderElements(), isEmpty);
+  });
+
+  test('overrideWithProvider preserves the state accross update', () async {
+    final provider = ChangeNotifierProvider((_) {
+      return TestNotifier();
+    });
+    final notifier = TestNotifier();
+    final notifier2 = TestNotifier();
+    final container = createContainer(overrides: [
+      provider.overrideWithProvider(ChangeNotifierProvider((_) => notifier)),
+    ]);
+    addTearDown(container.dispose);
+
+    var callCount = 0;
+    final sub = container.listen(provider, (prev, value) => callCount++);
+
+    expect(sub.read(), notifier);
+    expect(container.read(provider.notifier), notifier);
+    expect(notifier.hasListeners, true);
+    expect(callCount, 0);
+
+    notifier.count++;
+
+    await container.pump();
+    expect(callCount, 1);
+
+    container.updateOverrides([
+      provider.overrideWithProvider(ChangeNotifierProvider((_) => notifier2)),
+    ]);
+
+    await container.pump();
+    expect(callCount, 1);
+    expect(container.read(provider.notifier), notifier);
+    expect(notifier2.hasListeners, false);
+
+    notifier.count++;
+
+    await container.pump();
+    expect(callCount, 2);
+    expect(container.read(provider.notifier), notifier);
+    expect(notifier.mounted, true);
+
+    container.dispose();
+
+    expect(callCount, 2);
+    expect(notifier.mounted, false);
   });
 }