+++ packages/riverpod/lib/src/provider/base.dart
@@ -239,7 +239,9 @@ class ProviderElement<State> extends ProviderElementBase<State>
 /// {@macro riverpod.provider}
 @sealed
 class Provider<State> extends AlwaysAliveProviderBase<State>
-    with OverrideWithValueMixin<State> {
+    with
+        OverrideWithValueMixin<State>,
+        OverrideWithProviderMixin<State, AlwaysAliveProviderBase<State>> {
   /// {@macro riverpod.provider}
   Provider(
     this._create, {
@@ -252,6 +254,9 @@ class Provider<State> extends AlwaysAliveProviderBase<State>
 
   /// {@macro riverpod.autoDispose}
   static const autoDispose = AutoDisposeProviderBuilder();
+
+  @override
+  ProviderBase<State> get originProvider => this;
 
   @override
   final List<ProviderOrFamily>? dependencies;