+++ packages/riverpod/lib/src/state_notifier_provider/auto_dispose.dart
@@ -76,7 +76,10 @@ abstract class AutoDisposeStateNotifierProviderRef<
 @sealed
 class AutoDisposeStateNotifierProvider<Notifier extends StateNotifier<State>,
         State> extends AutoDisposeProviderBase<State>
-    with StateNotifierProviderOverrideMixin<Notifier, State> {
+    with
+        StateNotifierProviderOverrideMixin<Notifier, State>,
+        OverrideWithProviderMixin<Notifier,
+            AutoDisposeStateNotifierProvider<Notifier, State>> {
   /// {@macro riverpod.statenotifierprovider}
   AutoDisposeStateNotifierProvider(
     Create<Notifier, AutoDisposeStateNotifierProviderRef<Notifier, State>>
@@ -208,5 +211,20 @@ class AutoDisposeStateNotifierProviderFamily<
     final provider = call(argument);
     setup(origin: provider, override: provider);
     setup(origin: provider.notifier, override: provider.notifier);
+  }
+
+  /// {@macro riverpod.overridewithprovider}
+  Override overrideWithProvider(
+    AutoDisposeStateNotifierProvider<Notifier, State> Function(Arg argument)
+        override,
+  ) {
+    return FamilyOverride<Arg>(
+      this,
+      (arg, setup) {
+        final provider = call(arg);
+        setup(origin: provider.notifier, override: override(arg).notifier);
+        setup(origin: provider, override: provider);
+      },
+    );
   }
 }